# lib/api-zod/src/generated/types/finalizeUploadRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Sorrel - Modular Video Production Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface FinalizeUploadRequest {
  /**
     * The objectPath (or raw GCS upload URL) returned by request-url.
     * @minLength 1
     */
  objectPath: string;
}